Refactor to Coordinators
