+++ modulemate/src/main/java/at/xa1/modulemate/mode/ModeCoordinator.kt
@@ -2,6 +2,6 @@ package at.xa1.modulemate.mode
 
 import at.xa1.modulemate.ui.UiUserInput
 
-interface LiveUiMode {
-    fun print(input: UiUserInput?)
+interface ModeCoordinator {
+    fun run(): UiUserInput
 }
