+++ modulemate/src/main/java/at/xa1/modulemate/mode/command/CommandModeCoordinator.kt
@@ -3,17 +3,17 @@ package at.xa1.modulemate.mode.command
 import at.xa1.modulemate.UserCommandRunner
 import at.xa1.modulemate.command.Command
 import at.xa1.modulemate.command.CommandList
-import at.xa1.modulemate.mode.LiveUiMode
+import at.xa1.modulemate.mode.ModeCoordinator
 import at.xa1.modulemate.mode.SearchListScreen
 import at.xa1.modulemate.ui.ListItemRenderer
 import at.xa1.modulemate.ui.Ui
 import at.xa1.modulemate.ui.UiUserInput
 
-internal class CommandMode(
+internal class CommandModeCoordinator(
     private val ui: Ui,
     private val commandList: CommandList,
     private val commandRunner: UserCommandRunner
-) : LiveUiMode {
+) : ModeCoordinator {
     private val screen = SearchListScreen(
         emoji = "\uD83D\uDD79", // = 🕹️but without the  U+FE0F (VARIATION SELECTOR-16)
         hint = "Command Mode",
@@ -31,12 +31,12 @@ internal class CommandMode(
         }
     )
 
-    override fun print(input: UiUserInput?) {
+    override fun run(): UiUserInput {
         screen.print(ui)
 
         while (true) {
             when (val input = ui.readUserInput()) {
-                UiUserInput.Tab, UiUserInput.Shift.Tab -> return
+                UiUserInput.Tab, UiUserInput.Shift.Tab -> return input
                 UiUserInput.Return -> executeCommand()
                 else -> {
                     screen.input(input)
