+++ modulemate/src/main/java/at/xa1/modulemate/mode/help/HelpModeCoordinator.kt
@@ -2,14 +2,14 @@ package at.xa1.modulemate.mode.help
 
 import at.xa1.modulemate.Modulemate
 import at.xa1.modulemate.cli.CliColor
-import at.xa1.modulemate.mode.LiveUiMode
+import at.xa1.modulemate.mode.ModeCoordinator
 import at.xa1.modulemate.mode.SearchListScreen
 import at.xa1.modulemate.ui.Ui
 import at.xa1.modulemate.ui.UiUserInput
 
-internal class HelpMode(
+internal class HelpModeCoordinator(
     private val ui: Ui
-) : LiveUiMode {
+) : ModeCoordinator {
     private val screen = SearchListScreen(
         emoji = "ℹ\uFE0F",
         hint = "Help Mode",
@@ -44,20 +44,19 @@ internal class HelpMode(
         listItemRenderer = { item, _ -> " " + CliColor.RESET + " " + item }
     )
 
-    override fun print(input: UiUserInput?) {
+    override fun run(): UiUserInput {
         screen.print(ui)
 
         while (true) {
             when (val input = ui.readUserInput()) {
-                UiUserInput.Tab, UiUserInput.Shift.Tab -> return
+                UiUserInput.Tab, UiUserInput.Shift.Tab -> return input
                 UiUserInput.Arrow.Up,
-                UiUserInput.Arrow.Down
-                -> {
+                UiUserInput.Arrow.Down -> {
                     screen.input(input)
                     screen.print(ui)
                 }
 
-                else -> return
+                else -> return input
             }
         }
     }
